Address Round 3 feedback for saved-filter shape validation
The `saveFilter` server action only verified `filter.mode` at runtime;
a crafted authenticated client could persist a structured payload
whose inner `EventListFilterInput` carried a schema-violating field
(e.g. `keywords: "not-an-array"`). The row would then crash the chip
bar / drawer the next time the rail loaded it.

Lift the existing `coerceFilter` / `coerceEventListFilterInput`
helpers out of `filter-url.ts` into a shared
`src/lib/detection/filter-coerce.ts` module so the same runtime
shape boundary protects both the `?f=` URL decoder and the saved-
filter store. `storedPayloadFromFilter` now coerces before insert
(bad fields are dropped; an unrecoverable outer shape throws
`SavedFilterInvalidError`, which the action surfaces as the generic
`server-error` code), and `filterFromStoredPayload` coerces stored
`filter_json` on read for defense-in-depth against any pre-existing
row.

Part of #286
