+++ src/lib/detection/filter-url.ts
@@ -24,16 +24,9 @@
 
 import type { EndpointEntry } from "./endpoint-filter";
 import type { Filter } from "./filter";
+import { coerceFilter, coerceIpRange } from "./filter-coerce";
 import { PERIOD_KEYS, type PeriodKey } from "./period";
-import type {
-  EndpointInput,
-  EventListFilterInput,
-  FlowKind,
-  HostNetworkGroupInput,
-  IpRangeInput,
-  LearningMethod,
-  TrafficDirection,
-} from "./types";
+import type { EventListFilterInput } from "./types";
 import type { PivotFilterParams } from "./url-filters";
 
 /** URL search-param key for the encoded filter blob. */
@@ -208,156 +201,6 @@ function isFilter(value: unknown): value is Filter {
     return typeof (v as { text?: unknown }).text === "string";
   }
   return false;
-}
-
-/**
- * Reviewer Round 6 (item 2): coerce a parsed `?f=` payload's filter
- * field into a strict {@link Filter}. The shallow `isFilter` check
- * only verifies `mode` and that `input` is an object — it would
- * happily pass `{ categories: "not-an-array" }` straight through to
- * the chip / draft helpers, where `.filter(...)` on a string then
- * throws inside the render. The coercion below walks every
- * `EventListFilterInput` field and drops any that does not match the
- * generated schema's type. Returns `null` only when the outer shape
- * is unrecoverable (unknown mode, missing query text); a structured
- * filter with all fields stripped is still a valid filter.
- */
-function coerceFilter(value: unknown): Filter | null {
-  if (!value || typeof value !== "object") return null;
-  const v = value as { mode?: unknown; input?: unknown; text?: unknown };
-  if (v.mode === "structured") {
-    if (!v.input || typeof v.input !== "object") return null;
-    return {
-      mode: "structured",
-      input: coerceEventListFilterInput(v.input),
-    };
-  }
-  if (v.mode === "query") {
-    if (typeof v.text !== "string") return null;
-    return { mode: "query", text: v.text };
-  }
-  return null;
-}
-
-const FLOW_KIND_VALUES = new Set<FlowKind>(["INBOUND", "OUTBOUND", "INTERNAL"]);
-const LEARNING_METHOD_VALUES = new Set<LearningMethod>([
-  "UNSUPERVISED",
-  "SEMI_SUPERVISED",
-]);
-const TRAFFIC_DIRECTION_VALUES = new Set<TrafficDirection>(["FROM", "TO"]);
-
-function coerceEventListFilterInput(value: object): EventListFilterInput {
-  const v = value as Record<string, unknown>;
-  const out: EventListFilterInput = {};
-  if (typeof v.start === "string") out.start = v.start;
-  if (typeof v.end === "string") out.end = v.end;
-  if (typeof v.source === "string") out.source = v.source;
-  if (typeof v.destination === "string") out.destination = v.destination;
-  const stringArrayFields = [
-    "customers",
-    "keywords",
-    "networkTags",
-    "sensors",
-    "os",
-    "devices",
-    "hostnames",
-    "userIds",
-    "userNames",
-    "userDepartments",
-    "countries",
-    "kinds",
-    "triagePolicies",
-  ] as const;
-  for (const key of stringArrayFields) {
-    const arr = filterStringArray(v[key]);
-    if (arr) out[key] = arr;
-  }
-  const numberArrayFields = ["levels"] as const;
-  for (const key of numberArrayFields) {
-    const arr = filterNumberArray(v[key]);
-    if (arr) out[key] = arr;
-  }
-  if (Array.isArray(v.categories)) {
-    out.categories = v.categories.filter(
-      (item): item is number | null =>
-        item === null || typeof item === "number",
-    );
-  }
-  if (Array.isArray(v.directions)) {
-    out.directions = v.directions.filter(
-      (item): item is FlowKind =>
-        typeof item === "string" && FLOW_KIND_VALUES.has(item as FlowKind),
-    );
-  }
-  if (Array.isArray(v.learningMethods)) {
-    out.learningMethods = v.learningMethods.filter(
-      (item): item is LearningMethod =>
-        typeof item === "string" &&
-        LEARNING_METHOD_VALUES.has(item as LearningMethod),
-    );
-  }
-  if (typeof v.confidenceMin === "number" && Number.isFinite(v.confidenceMin)) {
-    out.confidenceMin = v.confidenceMin;
-  }
-  if (typeof v.confidenceMax === "number" && Number.isFinite(v.confidenceMax)) {
-    out.confidenceMax = v.confidenceMax;
-  }
-  if (Array.isArray(v.endpoints)) {
-    const eps = v.endpoints
-      .map(coerceEndpointInput)
-      .filter((ep): ep is EndpointInput => ep !== null);
-    if (eps.length > 0) out.endpoints = eps;
-  }
-  return out;
-}
-
-function filterStringArray(value: unknown): string[] | null {
-  if (!Array.isArray(value)) return null;
-  return value.filter((item): item is string => typeof item === "string");
-}
-
-function filterNumberArray(value: unknown): number[] | null {
-  if (!Array.isArray(value)) return null;
-  return value.filter(
-    (item): item is number => typeof item === "number" && Number.isFinite(item),
-  );
-}
-
-function coerceEndpointInput(value: unknown): EndpointInput | null {
-  if (!value || typeof value !== "object") return null;
-  const v = value as Record<string, unknown>;
-  const out: EndpointInput = {};
-  if (
-    typeof v.direction === "string" &&
-    TRAFFIC_DIRECTION_VALUES.has(v.direction as TrafficDirection)
-  ) {
-    out.direction = v.direction as TrafficDirection;
-  } else if (v.direction === null) {
-    out.direction = null;
-  }
-  if (typeof v.predefined === "string") out.predefined = v.predefined;
-  const custom = coerceHostNetworkGroup(v.custom);
-  if (custom) out.custom = custom;
-  return out;
-}
-
-function coerceHostNetworkGroup(value: unknown): HostNetworkGroupInput | null {
-  if (!value || typeof value !== "object") return null;
-  const v = value as Record<string, unknown>;
-  return {
-    hosts: filterStringArray(v.hosts) ?? [],
-    networks: filterStringArray(v.networks) ?? [],
-    ranges: Array.isArray(v.ranges)
-      ? v.ranges.map(coerceIpRange).filter((r): r is IpRangeInput => r !== null)
-      : [],
-  };
-}
-
-function coerceIpRange(value: unknown): IpRangeInput | null {
-  if (!value || typeof value !== "object") return null;
-  const v = value as Record<string, unknown>;
-  if (typeof v.start !== "string" || typeof v.end !== "string") return null;
-  return { start: v.start, end: v.end };
 }
 
 function coercePeriodKey(value: unknown): PeriodKey | null {
